completed app

+++ Controllers/LoadsController.cs
@@ -19,6 +19,15 @@ namespace CargoManagementAPI.Controllers
         [HttpGet]
         public ActionResult<LoadsDto> GetLoads([FromQuery] string pageCursor = "")
         {
+            // check that Accept response type is JSON or any
+            var acceptType = Request.Headers["Accept"];
+            if (!acceptType.Contains("application/json") && !acceptType.Contains("*/*"))
+            {
+                var error = new ErrorMessage("API does not return requests of type " + acceptType +
+                                             ". API can only return application/json");
+                return StatusCode(406, error);
+            }
+            
             var uriString = 
                 $"{$"https://{this.Request.Host}{this.Request.PathBase}{this.Request.Path}"}";
             var baseUri = $"https://{this.Request.Host}";
@@ -29,6 +38,15 @@ namespace CargoManagementAPI.Controllers
         [HttpGet("{loadId}")]
         public ActionResult<BoatDto> GetLoad(long loadId)
         {
+            // check that Accept response type is JSON or any
+            var acceptType = Request.Headers["Accept"];
+            if (!acceptType.Contains("application/json") && !acceptType.Contains("*/*"))
+            {
+                var error = new ErrorMessage("API does not return requests of type " + acceptType +
+                                             ". API can only return application/json");
+                return StatusCode(406, error);
+            }
+            
             var uriString =
                 $"{$"https://{this.Request.Host}{this.Request.PathBase}{this.Request.Path}"}";
             var baseUri = $"https://{this.Request.Host}";
@@ -46,6 +64,15 @@ namespace CargoManagementAPI.Controllers
         [HttpPost]
         public ActionResult CreateLoad([FromBody] LoadDto newLoad)
         {
+            // check that Accept response type is JSON or any
+            var acceptType = Request.Headers["Accept"];
+            if (!acceptType.Contains("application/json") && !acceptType.Contains("*/*"))
+            {
+                var error = new ErrorMessage("API does not return requests of type " + acceptType +
+                                             ". API can only return application/json");
+                return StatusCode(406, error);
+            }
+            
             if (newLoad.Volume == null || newLoad.Content == null)
             {
                 var error = new ErrorMessage("The request object is missing required volume or required content");
@@ -86,9 +113,50 @@ namespace CargoManagementAPI.Controllers
             return Ok(load);
         }
 
+        [HttpPut("{loadId}")]
+        public ActionResult EditLoad([FromBody] LoadDto editedLoad, long loadId)
+        {
+            // check that Accept response type is JSON or any
+            var acceptType = Request.Headers["Accept"];
+            if (!acceptType.Contains("application/json") && !acceptType.Contains("*/*"))
+            {
+                var error = new ErrorMessage("API does not return requests of type " + acceptType +
+                                             ". API can only return application/json");
+                return StatusCode(406, error);
+            }
+
+            if (editedLoad.Content == null || editedLoad.Volume == null)
+            {
+                var error = new ErrorMessage("The request object is missing at least one of the required attributes");
+                return BadRequest(error);
+            }
+
+            var uriString =
+                $"{$"https://{this.Request.Host}{this.Request.PathBase}{this.Request.Path}"}";
+
+            var load = query.EditLoadQuery(loadId, editedLoad, uriString);
+
+            if (load == null)
+            {
+                var error = new ErrorMessage("No boat with this boat_id exists or token provided is not owner of boat");
+                return NotFound(error);
+            }
+
+            return this.SeeOther(load.Self);
+        }
+
         [HttpDelete("{loadId}")]
         public ActionResult DeleteLoad(long loadId)
         {
+            // check that Accept response type is JSON or any
+            var acceptType = Request.Headers["Accept"];
+            if (!acceptType.Contains("application/json") && !acceptType.Contains("*/*"))
+            {
+                var error = new ErrorMessage("API does not return requests of type " + acceptType +
+                                             ". API can only return application/json");
+                return StatusCode(406, error);
+            }
+            
             var deleteSuccess = query.DeleteLoadQuery(loadId);
         
             if (!deleteSuccess)
